requirements.txt
pyzmq==22.3.0
requests==2.22.0
pytest==7.1.1